requirements.txt
Django==1.9.5
django-guardian==1.4.4
six==1.10.0
